requirements.txt
aiohttp==3.7.4.post0
asgiref==3.3.4
async-timeout==3.0.1
attrs==20.3.0
chardet==4.0.0
Django==3.0
djangorestframework==3.12.1
idna==3.1
idna-ssl==1.1.0
multidict==5.1.0
Pillow==7.2.0
psycopg2-binary==2.8.6
pytz==2021.1
sqlparse==0.4.1
telepot==12.7
typing-extensions==3.7.4.3
urllib3==1.26.4
yarl==1.6.3
